Refactor to component
